:recycle: Refactor: Part II.
