--- data/data.go
@@ -1,396 +1,0 @@
-package data
-
-import (
-	"database/sql"
-	"fmt"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
-)
-
-type Pulsar struct {
-	ID           int64
-	PSRJ         sql.NullString
-	PSRJ_REF     sql.NullString
-	RAJ          sql.NullString
-	RAJ_ERR      sql.NullFloat64
-	RAJ_REF      sql.NullString
-	DECJ         sql.NullString
-	DECJ_ERR     sql.NullFloat64
-	DECJ_REF     sql.NullString
-	POSEPOCH     sql.NullFloat64
-	POSEPOCH_REF sql.NullString
-	F0           sql.NullFloat64
-	F0_ERR       sql.NullFloat64
-	F0_REF       sql.NullString
-	F1           sql.NullFloat64
-	F1_ERR       sql.NullFloat64
-	F1_REF       sql.NullString
-	PEPOCH       sql.NullFloat64
-	PEPOCH_REF   sql.NullString
-	DM           sql.NullFloat64
-	DM_ERR       sql.NullFloat64
-	DM_REF       sql.NullString
-	S1400        sql.NullFloat64
-	S1400_REF    sql.NullString
-	DIST_DM      sql.NullFloat64
-	DIST_DM_REF  sql.NullString
-	DIST_DM1     sql.NullFloat64
-	DIST_DM1_REF sql.NullString
-	TYPE         sql.NullString
-	ASSOC        sql.NullString
-	SURVEY       sql.NullString
-	JNAME        sql.NullString
-	NAME         sql.NullString
-	JNAME_REF    sql.NullString
-	NAME_REF     sql.NullString
-	RAJD         sql.NullFloat64
-	DECJD        sql.NullFloat64
-	RAJD_ERR     sql.NullFloat64
-	DECJD_ERR    sql.NullFloat64
-	DATE         sql.NullFloat64
-	P0           sql.NullFloat64
-	P0_ERR       sql.NullFloat64
-	P0_REF       sql.NullString
-	P1           sql.NullFloat64
-	P1_ERR       sql.NullFloat64
-	P1_REF       sql.NullString
-	DMEPOCH      sql.NullFloat64
-	DMEPOCH_REF  sql.NullString
-	RM           sql.NullFloat64
-	RM_ERR       sql.NullFloat64
-	RM_REF       sql.NullString
-	S400         sql.NullFloat64
-	S400_REF     sql.NullString
-	S150         sql.NullFloat64
-	S150_ERR     sql.NullFloat64
-	S150_REF     sql.NullString
-	SPINDX       sql.NullFloat64
-	SPINDX_ERR   sql.NullFloat64
-	SPINDX_REF   sql.NullString
-	W50          sql.NullFloat64
-	W50_REF      sql.NullString
-	W10          sql.NullFloat64
-	W10_REF      sql.NullString
-	F2           sql.NullFloat64
-	F2_ERR       sql.NullFloat64
-	F2_REF       sql.NullString
-	F3           sql.NullFloat64
-	F3_ERR       sql.NullFloat64
-	F3_REF       sql.NullString
-	NGLT         sql.NullFloat64
-	DIST_AMN     sql.NullFloat64
-	DIST_AMN_REF sql.NullString
-	DIST_AMX     sql.NullFloat64
-	DIST_AMX_REF sql.NullString
-	S300         sql.NullFloat64
-	S300_REF     sql.NullString
-	PSRB         sql.NullString
-	PSRB_REF     sql.NullString
-	PMRA         sql.NullFloat64
-	PMRA_ERR     sql.NullFloat64
-	PMRA_REF     sql.NullString
-	PMDEC        sql.NullFloat64
-	PMDEC_ERR    sql.NullFloat64
-	PMDEC_REF    sql.NullString
-	S400_ERR     sql.NullFloat64
-	S50          sql.NullFloat64
-	S50_ERR      sql.NullFloat64
-	S50_REF      sql.NullString
-	S600         sql.NullFloat64
-	S600_ERR     sql.NullFloat64
-	S600_REF     sql.NullString
-	S1400_ERR    sql.NullFloat64
-	TAU_SC       sql.NullFloat64
-	TAU_SC_ERR   sql.NullFloat64
-	TAU_SC_REF   sql.NullString
-	BNAME        sql.NullString
-	BNAME_REF    sql.NullString
-	ELONG        sql.NullFloat64
-	ELONG_ERR    sql.NullFloat64
-	ELONG_REF    sql.NullString
-	ELAT         sql.NullFloat64
-	ELAT_ERR     sql.NullFloat64
-	ELAT_REF     sql.NullString
-	PMELONG      sql.NullFloat64
-	PMELONG_ERR  sql.NullFloat64
-	PMELONG_REF  sql.NullString
-	PMELAT       sql.NullFloat64
-	PMELAT_ERR   sql.NullFloat64
-	PMELAT_REF   sql.NullString
-	PX           sql.NullFloat64
-	PX_ERR       sql.NullFloat64
-	PX_REF       sql.NullString
-	UNITS        sql.NullString
-	UNITS_REF    sql.NullString
-	BINCOMP      sql.NullString
-	BINARY       sql.NullString
-	BINARY_REF   sql.NullString
-	TASC         sql.NullFloat64
-	TASC_ERR     sql.NullFloat64
-	TASC_REF     sql.NullString
-	FB0          sql.NullFloat64
-	FB0_ERR      sql.NullFloat64
-	FB0_REF      sql.NullString
-	FB1          sql.NullFloat64
-	FB1_ERR      sql.NullFloat64
-	FB1_REF      sql.NullString
-	FB2          sql.NullFloat64
-	FB2_ERR      sql.NullFloat64
-	FB2_REF      sql.NullString
-	A1           sql.NullFloat64
-	A1_ERR       sql.NullFloat64
-	A1_REF       sql.NullString
-	EPS1         sql.NullFloat64
-	EPS1_ERR     sql.NullFloat64
-	EPS1_REF     sql.NullString
-	EPS2         sql.NullFloat64
-	EPS2_ERR     sql.NullFloat64
-	EPS2_REF     sql.NullString
-	CLK          sql.NullString
-	CLK_REF      sql.NullString
-	EPHEM        sql.NullString
-	EPHEM_REF    sql.NullString
-	PB           sql.NullFloat64
-	PB_REF       sql.NullString
-	DIST_A       sql.NullFloat64
-	DIST_A_ERR   sql.NullFloat64
-	DIST_A_REF   sql.NullString
-	S3000        sql.NullFloat64
-	S3000_ERR    sql.NullFloat64
-	S3000_REF    sql.NullString
-	S4000        sql.NullFloat64
-	S4000_ERR    sql.NullFloat64
-	S4000_REF    sql.NullString
-	S800         sql.NullFloat64
-	S800_ERR     sql.NullFloat64
-	S800_REF     sql.NullString
-	S2000        sql.NullFloat64
-	S2000_ERR    sql.NullFloat64
-	S2000_REF    sql.NullString
-	S900         sql.NullFloat64
-	S900_ERR     sql.NullFloat64
-	S900_REF     sql.NullString
-	S1600        sql.NullFloat64
-	S1600_ERR    sql.NullFloat64
-	S1600_REF    sql.NullString
-	T0           sql.NullFloat64
-	T0_ERR       sql.NullFloat64
-	T0_REF       sql.NullString
-	PB_ERR       sql.NullFloat64
-	PBDOT        sql.NullFloat64
-	PBDOT_ERR    sql.NullFloat64
-	PBDOT_REF    sql.NullString
-	ECC          sql.NullFloat64
-	ECC_ERR      sql.NullFloat64
-	ECC_REF      sql.NullString
-	OM           sql.NullFloat64
-	OM_ERR       sql.NullFloat64
-	OM_REF       sql.NullString
-	OMDOT        sql.NullFloat64
-	OMDOT_ERR    sql.NullFloat64
-	OMDOT_REF    sql.NullString
-	M2           sql.NullFloat64
-	M2_ERR       sql.NullFloat64
-	M2_REF       sql.NullString
-	S350         sql.NullFloat64
-	S350_ERR     sql.NullFloat64
-	S350_REF     sql.NullString
-	W50_ERR      sql.NullFloat64
-	W10_ERR      sql.NullFloat64
-	S300_ERR     sql.NullFloat64
-	S100         sql.NullFloat64
-	S100_REF     sql.NullString
-	S200         sql.NullFloat64
-	S200_ERR     sql.NullFloat64
-	S200_REF     sql.NullString
-	S80          sql.NullFloat64
-	S80_ERR      sql.NullFloat64
-	S80_REF      sql.NullString
-	A1DOT        sql.NullFloat64
-	A1DOT_ERR    sql.NullFloat64
-	A1DOT_REF    sql.NullString
-	S60          sql.NullFloat64
-	S60_ERR      sql.NullFloat64
-	S60_REF      sql.NullString
-	S5000        sql.NullFloat64
-	S5000_ERR    sql.NullFloat64
-	S5000_REF    sql.NullString
-	DM1          sql.NullFloat64
-	DM1_ERR      sql.NullFloat64
-	DM1_REF      sql.NullString
-	S100_ERR     sql.NullFloat64
-	S40          sql.NullFloat64
-	S40_ERR      sql.NullFloat64
-	S40_REF      sql.NullString
-	TASC_2       sql.NullFloat64
-	TASC_2_ERR   sql.NullFloat64
-	TASC_2_REF   sql.NullString
-	PB_2         sql.NullFloat64
-	PB_2_ERR     sql.NullFloat64
-	PB_2_REF     sql.NullString
-	A1_2         sql.NullFloat64
-	A1_2_ERR     sql.NullFloat64
-	A1_2_REF     sql.NullString
-	EPS1_2       sql.NullFloat64
-	EPS1_2_ERR   sql.NullFloat64
-	EPS1_2_REF   sql.NullString
-	EPS2_2       sql.NullFloat64
-	EPS2_2_ERR   sql.NullFloat64
-	EPS2_2_REF   sql.NullString
-	SINI         sql.NullFloat64
-	SINI_ERR     sql.NullFloat64
-	SINI_REF     sql.NullString
-	M2_2         sql.NullFloat64
-	M2_2_ERR     sql.NullFloat64
-	M2_2_REF     sql.NullString
-	MASS_Q       sql.NullFloat64
-	MASS_Q_ERR   sql.NullFloat64
-	MASS_Q_REF   sql.NullString
-	S700         sql.NullFloat64
-	S700_ERR     sql.NullFloat64
-	S700_REF     sql.NullString
-	KOM          sql.NullFloat64
-	KOM_ERR      sql.NullFloat64
-	KOM_REF      sql.NullString
-	KIN          sql.NullFloat64
-	KIN_ERR      sql.NullFloat64
-	KIN_REF      sql.NullString
-	H3           sql.NullFloat64
-	H3_ERR       sql.NullFloat64
-	H3_REF       sql.NullString
-	STIG         sql.NullFloat64
-	STIG_ERR     sql.NullFloat64
-	STIG_REF     sql.NullString
-	H4           sql.NullFloat64
-	H4_ERR       sql.NullFloat64
-	H4_REF       sql.NullString
-	GAMMA        sql.NullFloat64
-	GAMMA_ERR    sql.NullFloat64
-	GAMMA_REF    sql.NullString
-	DM2          sql.NullFloat64
-	DM2_ERR      sql.NullFloat64
-	DM2_REF      sql.NullString
-	S100G        sql.NullFloat64
-	S100G_ERR    sql.NullFloat64
-	S100G_REF    sql.NullString
-	S150G        sql.NullFloat64
-	S150G_ERR    sql.NullFloat64
-	S150G_REF    sql.NullString
-	OM_ASC       sql.NullFloat64
-	OM_ASC_ERR   sql.NullFloat64
-	OM_ASC_REF   sql.NullString
-	DM3          sql.NullFloat64
-	DM3_ERR      sql.NullFloat64
-	DM3_REF      sql.NullString
-	T0_2         sql.NullFloat64
-	T0_2_ERR     sql.NullFloat64
-	T0_2_REF     sql.NullString
-	T0_3         sql.NullFloat64
-	T0_3_ERR     sql.NullFloat64
-	T0_3_REF     sql.NullString
-	PB_3         sql.NullFloat64
-	PB_3_ERR     sql.NullFloat64
-	PB_3_REF     sql.NullString
-	A1_3         sql.NullFloat64
-	A1_3_ERR     sql.NullFloat64
-	A1_3_REF     sql.NullString
-	ECC_2        sql.NullFloat64
-	ECC_2_ERR    sql.NullFloat64
-	ECC_2_REF    sql.NullString
-	ECC_3        sql.NullFloat64
-	ECC_3_ERR    sql.NullFloat64
-	ECC_3_REF    sql.NullString
-	OM_2         sql.NullFloat64
-	OM_2_ERR     sql.NullFloat64
-	OM_2_REF     sql.NullString
-	OM_3         sql.NullFloat64
-	OM_3_ERR     sql.NullFloat64
-	OM_3_REF     sql.NullString
-	SINI_2       sql.NullFloat64
-	SINI_2_ERR   sql.NullFloat64
-	SINI_2_REF   sql.NullString
-	SINI_3       sql.NullFloat64
-	SINI_3_ERR   sql.NullFloat64
-	SINI_3_REF   sql.NullString
-	M2_3         sql.NullFloat64
-	M2_3_ERR     sql.NullFloat64
-	M2_3_REF     sql.NullString
-	A12DOT       sql.NullFloat64
-	A12DOT_ERR   sql.NullFloat64
-	A12DOT_REF   sql.NullString
-	OM2DOT       sql.NullFloat64
-	OM2DOT_ERR   sql.NullFloat64
-	OM2DOT_REF   sql.NullString
-	S6000        sql.NullFloat64
-	S6000_ERR    sql.NullFloat64
-	S6000_REF    sql.NullString
-	XDOT         sql.NullFloat64
-	XDOT_ERR     sql.NullFloat64
-	XDOT_REF     sql.NullString
-	F4           sql.NullFloat64
-	F4_ERR       sql.NullFloat64
-	F4_REF       sql.NullString
-	F5           sql.NullFloat64
-	F5_ERR       sql.NullFloat64
-	F5_REF       sql.NullString
-	BINCOMP_REF  sql.NullString
-	S8000        sql.NullFloat64
-	S8000_ERR    sql.NullFloat64
-	S8000_REF    sql.NullString
-	S9000        sql.NullFloat64
-	S9000_REF    sql.NullString
-	S9000_ERR    sql.NullFloat64
-	ECCDOT       sql.NullFloat64
-	ECCDOT_ERR   sql.NullFloat64
-	ECCDOT_REF   sql.NullString
-	DTHETA       sql.NullFloat64
-	DTHETA_ERR   sql.NullFloat64
-	DTHETA_REF   sql.NullString
-	S1500        sql.NullFloat64
-	S1500_REF    sql.NullString
-	DIST         sql.NullFloat64
-	DIST1        sql.NullFloat64
-	GL           sql.NullFloat64
-	GB           sql.NullFloat64
-	XX           sql.NullFloat64
-	YY           sql.NullFloat64
-	ZZ           sql.NullFloat64
-	DMSINB       sql.NullFloat64
-	PMB          sql.NullFloat64
-	PML          sql.NullFloat64
-	AGE          sql.NullFloat64
-	BSURF        sql.NullFloat64
-	B_LC         sql.NullFloat64
-	EDOT         sql.NullFloat64
-	EDOTD2       sql.NullFloat64
-	PMTOT        sql.NullFloat64
-	PMTOT_ERR    sql.NullFloat64
-	VTRANS       sql.NullFloat64
-	P1_I         sql.NullFloat64
-	AGE_I        sql.NullFloat64
-	BSURF_I      sql.NullFloat64
-	EDOT_I       sql.NullFloat64
-	SI414        sql.NullFloat64
-	R_LUM        sql.NullFloat64
-	R_LUM14      sql.NullFloat64
-	MASSFN       sql.NullFloat64
-	MINMASS      sql.NullFloat64
-	MEDMASS      sql.NullFloat64
-	UPRMASS      sql.NullFloat64
-	MASSFN_ERR   sql.NullFloat64
-	MINOMDOT     sql.NullFloat64
-	H0_SD        sql.NullFloat64
-	ASSOC_ORIG   sql.NullString
-	ASSOC_REF    sql.NullString
-	TYPE_ORIG    sql.NullString
-	TYPE_REF     sql.NullString
-	BINCOMP_ORIG sql.NullString
-}
-
-func GetPulsars(jname string) []Pulsar {
-	db, _ := gorm.Open(sqlite.Open("./data/pulsars.db"), &gorm.Config{Logger: logger.Default.LogMode(logger.Silent)})
-	Pulsars := make([]Pulsar, 0)
-	db.Where("JNAME LIKE ?", fmt.Sprintf("%%%s%%", jname)).Find(&Pulsars)
-	return Pulsars
-}
